+++ NEWS.md
@@ -7,6 +7,7 @@
 * Small fix to support deprecated `field_occurrence` field in `validate` for convenience.
 * :bug: fix that prevented `BART` models to be saved/loaded from disk #127.
 * :bug: fixes related to `factor` handling for all engines.
+* :bug: fixes related to `is_comparable_raster` and `add_predictors`/`add_predictors_range`
 
 # ibis.iSDM 0.1.3